snes audio: fix BGM restarting every ~2.7s (32-bit phase overflow)
The mixer's phase accumulator was uint32 (16.16 fixed point), so idx = phase>>16
could only address 65536 source frames. The home BGM is 1256715 frames (52s), so
phase overflowed after 65536 frames (~2.7s at 24kHz) and wrapped idx back to 0 -
restarting the BGM from the top over and over (the reported 'loops the first few
seconds of the background music'). The loop_end check (1242352) never even ran.
Widen phase to uint64 so idx addresses the full sample; verified the BGM now
advances past frame 65536 and will loop at loop_end instead. SFX (all <65536
frames) were unaffected but are correct now too.

+++ emu/snes/snes_audio.c
@@ -49,7 +49,7 @@ void snes_audio_mix(snes_mixer *mx, int16_t *out, unsigned frames)
 		int acc = 0;
 		for (i = 0; i < SNES_AUD_VOICES; i++) {
 			snes_voice *v = &mx->v[i];
-			uint32_t idx;
+			uint64_t idx;
 			int s;
 			if (!v->active) continue;
 			idx = v->phase >> 16;
@@ -60,7 +60,7 @@ void snes_audio_mix(snes_mixer *mx, int16_t *out, unsigned frames)
 			idx = v->phase >> 16;
 			if (v->loop) {
 				if (idx >= v->loop_end)
-					v->phase -= (uint32_t)(v->loop_end - v->loop_start) << 16;
+					v->phase -= (uint64_t)(v->loop_end - v->loop_start) << 16;
 			} else if (idx >= v->len) {
 				v->active = 0;
 			}
